embeddinggemma_rag.py
#!/usr/bin/env python3
"""
Bengali Legal RAG with EmbeddingGemma
====================================

Clean, production-ready RAG system using pre-trained EmbeddingGemma-300M.
No fine-tuning required - achieves 76.4% accuracy on validation data.

FAISS Index: IndexFlatIP (optimal for EmbeddingGemma inner product search)
"""

import os
import pandas as pd
import numpy as np
from pathlib import Path
from typing import List, Dict
import logging
import re
import torch
import faiss
from sentence_transformers import SentenceTransformer
from sklearn.metrics import confusion_matrix, accuracy_score
import matplotlib.pyplot as plt
import seaborn as sns

# Environment setup
os.environ["TOKENIZERS_PARALLELISM"] = "false"
os.environ["OMP_NUM_THREADS"] = "1"

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class EmbeddingGemmaRAG:
    """Bengali Legal RAG System with EmbeddingGemma-300M"""
    
    def __init__(self, data_dir: str = "namjari_questions", confidence_threshold: float = 0.5):
        logger.info("🚀 Loading EmbeddingGemma-300M...")
        
        self.model = SentenceTransformer("google/embeddinggemma-300m")
        self.confidence_threshold = confidence_threshold
        self.data_dir = Path(data_dir)
        
        # Apple Silicon optimization
        if torch.backends.mps.is_available():
            torch.mps.empty_cache()
        
        # Data storage
        self.train_questions = []
        self.train_tags = []
        self.val_questions = []
        self.val_tags = []
        self.index = None
        
        # Build system
        self._create_validation_splits()
        self._load_training_data()
        self._build_faiss_index()
        self._evaluate_and_visualize()
    
    def _clean_text(self, text: str) -> str:
        """Basic Bengali text cleaning"""
        text = re.sub(r'[^\u0980-\u09FF\s\w]', '', text)
        return ' '.join(text.split()).strip()
    
    def _create_validation_splits(self):
        """Create validation splits: last 10 from each CSV"""
        logger.info("Creating validation splits (last 10 from each CSV)...")
        
        val_dir = Path("validation_data")
        val_dir.mkdir(exist_ok=True)
        
        for csv_file in self.data_dir.glob("*.csv"):
            df = pd.read_csv(csv_file)
            tag = csv_file.stem
            questions = df['question'].dropna().tolist()
            
            if len(questions) > 10:
                val_questions = questions[-10:]
                val_df = pd.DataFrame({'question': val_questions, 'tag': [tag] * len(val_questions)})
                val_df.to_csv(val_dir / f"{tag}_validation.csv", index=False)
                
                self.val_questions.extend([self._clean_text(q) for q in val_questions])
                self.val_tags.extend([tag] * len(val_questions))
        
        logger.info(f"✅ Validation set: {len(self.val_questions)} samples")
    
    def _load_training_data(self):
        """Load training data (excluding last 10 from each CSV)"""
        logger.info("Loading training data...")
        
        for csv_file in self.data_dir.glob("*.csv"):
            df = pd.read_csv(csv_file)
            tag = csv_file.stem
            questions = df['question'].dropna().tolist()
            
            # Take all except last 10 for training
            train_questions = questions[:-10] if len(questions) > 10 else questions
            train_questions = [self._clean_text(q) for q in train_questions]
            
            self.train_questions.extend(train_questions)
            self.train_tags.extend([tag] * len(train_questions))
        
        logger.info(f"✅ Training set: {len(self.train_questions)} samples")
    
    def _build_faiss_index(self):
        """Build FAISS IndexFlatIP for EmbeddingGemma"""
        logger.info("Building FAISS IndexFlatIP with EmbeddingGemma...")
        
        # Create embeddings with Classification prompt
        embeddings = self.model.encode(
            self.train_questions,
            prompt_name="Classification",
            show_progress_bar=True,
            batch_size=32,
            convert_to_numpy=True
        )
        
        # Build FAISS IndexFlatIP (optimal for EmbeddingGemma)
        dimension = embeddings.shape[1]
        self.index = faiss.IndexFlatIP(dimension)
        
        # Normalize for cosine similarity
        embeddings_norm = embeddings.astype('float32')
        faiss.normalize_L2(embeddings_norm)
        self.index.add(embeddings_norm)
        
        logger.info(f"✅ FAISS IndexFlatIP ready: {self.index.ntotal} vectors")
    
    def _evaluate_and_visualize(self):
        """Evaluate on validation data and create confusion matrix"""
        logger.info(f"Evaluating on {len(self.val_questions)} validation samples...")
        
        predictions = []
        confidences = []
        
        for val_question in self.val_questions:
            result = self.classify(val_question)
            predictions.append(result['predicted_tag'])
            confidences.append(result['confidence'])
        
        # Calculate metrics
        accuracy = accuracy_score(self.val_tags, predictions)
        avg_confidence = np.mean(confidences)
        confident_count = sum(1 for c in confidences if c >= self.confidence_threshold)
        
        logger.info(f"📊 Results:")
        logger.info(f"  Accuracy: {accuracy:.1%}")
        logger.info(f"  Confident Predictions: {confident_count}/{len(predictions)}")
        logger.info(f"  Average Confidence: {avg_confidence:.3f}")
        
        # Create confusion matrix
        unique_tags = sorted(list(set(self.val_tags + predictions)))
        cm = confusion_matrix(self.val_tags, predictions, labels=unique_tags)
        
        self._create_confusion_matrix(cm, unique_tags, accuracy, confident_count, len(predictions))
    
    def _create_confusion_matrix(self, cm, labels, accuracy, confident_count, total_count):
        """Create and save confusion matrix"""
        plt.figure(figsize=(14, 12))
        
        display_labels = [tag.replace('namjari_', '') for tag in labels]
        
        sns.heatmap(
            cm,
            annot=True,
            fmt='d',
            cmap='Blues',
            xticklabels=display_labels,
            yticklabels=display_labels,
            cbar_kws={'label': 'Count'}
        )
        
        plt.title(
            f'EmbeddingGemma Bengali Legal Classification\n'
            f'Accuracy: {accuracy:.1%} | FAISS IndexFlatIP | '
            f'Confident: {confident_count}/{total_count}',
            fontsize=16,
            pad=20
        )
        
        plt.xlabel('Predicted Category', fontsize=12)
        plt.ylabel('True Category', fontsize=12)
        plt.xticks(rotation=45, ha='right')
        plt.tight_layout()
        
        plt.savefig("confusion_matrix.png", dpi=300, bbox_inches='tight')
        plt.close()
        
        logger.info(f"✅ Confusion matrix saved: confusion_matrix.png")
    
    def classify(self, query: str, top_k: int = 5) -> Dict:
        """
        Classify Bengali legal query using semantic search
        
        Args:
            query: Bengali text to classify
            top_k: Number of similar examples to consider
            
        Returns:
            Classification result with confidence score
        """
        # Clean and encode query
        cleaned_query = self._clean_text(query)
        query_embedding = self.model.encode([cleaned_query], prompt_name="Classification")
        query_embedding = query_embedding.astype('float32')
        faiss.normalize_L2(query_embedding)
        
        # Search FAISS index
        similarities, indices = self.index.search(query_embedding, top_k)
        
        # Aggregate tag scores
        tag_scores = {}
        similar_examples = []
        
        for i in range(top_k):
            idx = indices[0][i]
            similarity = float(similarities[0][i])
            tag = self.train_tags[idx]
            question = self.train_questions[idx]
            
            similar_examples.append({
                'question': question,
                'tag': tag,
                'similarity': similarity,
                'rank': i + 1
            })
            
            # Weight by inverse rank
            weight = 1.0 / (i + 1)
            tag_scores[tag] = tag_scores.get(tag, 0) + similarity * weight
        
        # Best prediction
        best_tag = max(tag_scores.keys(), key=lambda t: tag_scores[t])
        
        # Normalize confidence
        max_possible = sum(1.0 / i for i in range(1, top_k + 1))
        confidence = min(1.0, tag_scores[best_tag] / max_possible)
        
        return {
            'query': query,
            'predicted_tag': best_tag,
            'confidence': confidence,
            'is_confident': confidence >= self.confidence_threshold,
            'tag_scores': tag_scores,
            'similar_questions': similar_examples
        }
    
    def get_stats(self) -> Dict:
        """Get system statistics"""
        return {
            'model': 'google/embeddinggemma-300m',
            'faiss_index': 'IndexFlatIP',
            'training_samples': len(self.train_questions),
            'validation_samples': len(self.val_questions),
            'total_tags': len(set(self.train_tags)),
            'confidence_threshold': self.confidence_threshold,
            'device': str(self.model.device)
        }


def main():
    """Main demo function"""
    print("🚀 EmbeddingGemma Bengali Legal RAG System")
    print("=" * 50)
    print("📋 Pre-trained EmbeddingGemma-300M + FAISS IndexFlatIP")
    print("=" * 50)
    
    # Initialize system (auto-creates validation and confusion matrix)
    rag = EmbeddingGemmaRAG()
    
    # Test sample queries  
    print(f"\n📝 Testing Sample Queries:")
    print("-" * 40)
    
    test_queries = [
        "আমি নামজারি করতে কত টাকা লাগবে?",
        "নামজারির জন্য কি কি কাগজপত্র লাগে?",
        "আমি নামজারি আবেদন কিভাবে করবো?"
    ]
    
    for query in test_queries:
        result = rag.classify(query)
        status = "✅" if result['is_confident'] else "⚠️"
        
        print(f"{status} {query}")
        print(f"   → {result['predicted_tag'].replace('namjari_', '')} (confidence: {result['confidence']:.3f})")
    
    # Show system stats
    stats = rag.get_stats()
    print(f"\n📊 System Statistics:")
    print(f"  Model: {stats['model']}")
    print(f"  FAISS Index: {stats['faiss_index']}")
    print(f"  Training Samples: {stats['training_samples']}")
    print(f"  Validation Samples: {stats['validation_samples']}")
    print(f"  Tags: {stats['total_tags']}")
    print(f"  Threshold: {stats['confidence_threshold']}")
    
    print(f"\n✅ System ready! Confusion matrix: confusion_matrix.png")
    
    return rag


if __name__ == "__main__":
    main()
