requirements.txt
# EmbeddingGemma Bengali Legal Classifier - Production Requirements
# Minimal dependencies for clean implementation

# Core ML libraries  
sentence-transformers>=5.0.0
torch>=2.0.0
transformers>=4.56.0

# FAISS for semantic search
faiss-cpu>=1.7.4

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Evaluation
scikit-learn>=1.1.0
matplotlib>=3.6.0
seaborn>=0.12.0